Added compressioned binary tree generator
- This saves space by sorting in x cordinate order

- Canvas only draws circles by given order which is faster

- Refactored code to input modules and strategies
