+++ treeVis/src/modules/input/strategy/index.ts
@@ -1,7 +1,10 @@
-export interface InputStrategy {
+import { Result } from "../../types";
+
+export interface InputStrategy<T> {
   draw(): void;
   setHeading(): void;
   setVisual(): void;
+  isValidInput(input?: string): Result<T>;
   validate(): void;
   format(isSuccess?: boolean, message?: string | null): void;
   initialize(): void;
