+++ treeVis/src/modules/storing/index.ts
@@ -1,0 +1,7 @@
+export function setLocalStorageItem(key: string, value: string) {
+  window.localStorage.setItem(key, value);
+}
+
+export function getLocalStorageItem(key: string) {
+  return window.localStorage.getItem(key);
+}
